requirements.txt
aiogram==3.25.0
anthropic==0.84.0
openai==2.21.0
google-generativeai==0.8.6
google-genai==1.63.0
google-api-python-client==2.190.0
google-auth==2.49.0.dev0
langchain==1.2.10
langchain-openai==1.1.10
langchain-google-genai==4.2.0
langchain-chroma==1.1.0
langchain-text-splitters==1.1.0
langchain-core==1.2.13
langchain-community==0.4.1
chromadb==1.5.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1
pandas==3.0.1
openpyxl==3.1.5
pypdf==6.7.1
python-docx==1.2.0
python-pptx==1.0.2
Flask==3.1.3
Pillow==12.1.1
tiktoken==0.12.0
